+++ src/locales/pl-PL.ts
@@ -53,6 +53,12 @@ export default {
   'menu.reports': 'Raporty',
   'menu.User Groups': 'Grupy',
   'menu.Templates': 'Szablony',
+  'menu.Settings': 'Ustawienia',
+  user_settings: 'Ustawienia użytkownika',
+  'user.settings': 'Ustawienia użytkownika',
+  settings: 'Settings',
+  package: 'Pakiet',
+  readonly: 'Tylko do odczytu',
   access: 'Dostęp',
   user_groups: 'Grupy',
   user_group: 'Grupa',
