+++ src/pages/Settings/index.tsx
@@ -1,245 +1,83 @@
-import { PlusOutlined } from '@ant-design/icons';
-import { Button, message, Tag, Tooltip, Popconfirm } from 'antd';
-import React, { useState, useRef } from 'react';
-import { useIntl, FormattedMessage } from 'umi';
+import { useMemo, useState, useEffect } from 'react';
+import { message, Spin, Row, Col } from 'antd';
+
+import ProCard from '@ant-design/pro-card';
+import { useParams, history } from 'umi';
 import { PageContainer } from '@ant-design/pro-layout';
-import type { IntlShape } from 'react-intl';
-import type { ProColumns, ActionType } from '@ant-design/pro-table';
-import ProTable from '@ant-design/pro-table';
+import { useIntl, FormattedMessage } from 'umi';
 
-import {
-  settings,
-  createSettings,
-  updateSettings,
-  deleteSettings,
-  settingGroups,
-} from '@/services/escola-lms/settings';
+import { configs as fetchConfigs } from '@/services/escola-lms/settings';
 
-import SettingsModalForm from './components/ModalForm';
-import { DeleteOutlined, EditOutlined } from '@ant-design/icons';
-import { useEffect } from 'react';
+import UserSettings from './user';
+import PackageForm from './package';
 
-const handleUpdate = async (intl: IntlShape, fields: API.Setting, id?: number) => {
-  const hide = message.loading(
-    intl.formatMessage({
-      id: 'loading',
-    }),
-  );
-  try {
-    await (id && id !== -1 ? updateSettings(id, { ...fields }) : createSettings({ ...fields }));
-    hide();
-    message.success(
-      intl.formatMessage({
-        id: 'success',
-      }),
-    );
-    return true;
-  } catch (error) {
-    hide();
-    message.error(
-      intl.formatMessage({
-        id: 'error',
-      }),
-    );
-    return false;
-  }
-};
+const sanitizePackageName = (name: string) => name.replaceAll('escolalms', '').split('_').join('');
 
-const handleRemove = async (intl: IntlShape, id: number) => {
-  const hide = message.loading(
-    intl.formatMessage({
-      id: 'loading',
-    }),
-  );
-  try {
-    await deleteSettings(id);
-    hide();
-    message.success(
-      intl.formatMessage({
-        id: 'success',
-      }),
-    );
-    return true;
-  } catch (error) {
-    hide();
-    message.error(
-      intl.formatMessage({
-        id: 'error',
-      }),
-    );
-    return false;
-  }
-};
-
-const TableList: React.FC = () => {
-  const [modalVisible, setModalVisible] = useState<number | false>(false);
-
-  const [groups, setGroups] = useState<string[]>([]);
-
-  const actionRef = useRef<ActionType>();
+export default () => {
+  const params = useParams<{ course?: string; tab?: string }>();
   const intl = useIntl();
+  const { tab = 'user_settings' } = params;
+
+  const [configs, setConfigs] = useState<API.Configs>();
 
   useEffect(() => {
-    settingGroups().then((response) => {
+    fetchConfigs().then((response) => {
       if (response.success) {
-        setGroups(response.data);
+        setConfigs(response.data);
       }
     });
   }, []);
 
-  const columns: ProColumns<API.Setting>[] = [
-    {
-      title: <FormattedMessage id="group" defaultMessage="group" />,
-      dataIndex: 'group',
-      hideInSearch: false,
-      key: 'group',
-      valueType: 'select',
-      width: 100,
-      valueEnum: groups.sort().reduce((acc, group) => ({ ...acc, [group]: group }), {}),
-    },
-    {
-      title: <FormattedMessage id="key" defaultMessage="key" />,
-      dataIndex: 'key',
-      hideInSearch: true,
-    },
-    {
-      title: <FormattedMessage id="type" defaultMessage="type" />,
-      dataIndex: 'type',
-      hideInSearch: true,
-    },
-
-    {
-      title: <FormattedMessage id="enumerable" defaultMessage="enumerable" />,
-      dataIndex: 'enumerable',
-      hideInForm: true,
-      hideInSearch: true,
-      render: (_, record) => (
-        <Tag color={record.enumerable ? 'success' : 'error'}>
-          <FormattedMessage
-            id={record.enumerable ? 'true' : 'false'}
-            defaultMessage={record.enumerable ? 'true' : 'false'}
-          />
-        </Tag>
-      ),
-    },
-
-    {
-      title: <FormattedMessage id="public" defaultMessage="public" />,
-      dataIndex: 'public',
-      hideInForm: true,
-      hideInSearch: true,
-      render: (_, record) => (
-        <Tag color={record.public ? 'success' : 'error'}>
-          <FormattedMessage
-            id={record.public ? 'true' : 'false'}
-            defaultMessage={record.public ? 'true' : 'false'}
-          />
-        </Tag>
-      ),
-    },
-
-    {
-      hideInSearch: true,
-      title: <FormattedMessage id="pages.searchTable.titleOption" defaultMessage="操作" />,
-      dataIndex: 'option',
-      valueType: 'option',
-      render: (_, record) => [
-        <Tooltip key="edit" title={<FormattedMessage id="edit" defaultMessage="edit" />}>
-          <Button
-            type="primary"
-            icon={<EditOutlined />}
-            onClick={() => setModalVisible(record.id)}
-          ></Button>
-        </Tooltip>,
-
-        <Popconfirm
-          key="delete"
-          title={
-            <FormattedMessage
-              id="deleteQuestion"
-              defaultMessage="Are you sure to delete this record?"
-            />
-          }
-          onConfirm={async () => {
-            const success = await handleRemove(intl, record.id);
-            if (success) {
-              setModalVisible(false);
-              if (actionRef.current) {
-                actionRef.current.reload();
-              }
-            }
-          }}
-          okText={<FormattedMessage id="yes" defaultMessage="Yes" />}
-          cancelText={<FormattedMessage id="no" defaultMessage="No" />}
-        >
-          <Tooltip title={<FormattedMessage id="delete" defaultMessage="delete" />}>
-            <Button type="primary" icon={<DeleteOutlined />} danger></Button>
-          </Tooltip>
-        </Popconfirm>,
-      ],
-    },
-  ];
+  if (!configs) {
+    return <Spin />;
+  }
 
   return (
-    <PageContainer>
-      <ProTable<API.Setting, API.PageParams & { group: string }>
-        headerTitle={intl.formatMessage({
-          id: 'menu.settings',
-        })}
-        actionRef={actionRef}
-        rowKey="id"
-        search={{
-          labelWidth: 120,
-        }}
-        toolBarRender={() => [
-          <Button
-            type="primary"
-            key="primary"
-            onClick={() => {
-              setModalVisible(-1);
-            }}
-          >
-            <PlusOutlined /> <FormattedMessage id="pages.searchTable.new" />
-          </Button>,
-        ]}
-        request={({ pageSize, current, group }) => {
-          return settings({
-            pageSize,
-            current,
-            group,
-          }).then((response) => {
-            if (response.success) {
-              return {
-                data: response.data,
-                total: response.meta.total,
-                success: true,
-              };
-            }
-            return [];
-          });
-        }}
-        columns={columns}
-      />
+    <PageContainer
+      title={<FormattedMessage id={'settings'} />}
+      header={{
+        breadcrumb: {
+          routes: [
+            {
+              path: 'settings',
+              breadcrumbName: intl.formatMessage({
+                id: 'menu.Settings',
+              }),
+            },
 
-      <SettingsModalForm
-        groups={groups}
-        id={modalVisible}
-        visible={Number.isInteger(modalVisible)}
-        onVisibleChange={(value) => {
-          return value === false && setModalVisible(false);
+            {
+              path: String(tab),
+              breadcrumbName: String(tab),
+            },
+          ],
+        },
+      }}
+    >
+      <ProCard
+        tabs={{
+          type: 'card',
+          activeKey: tab,
+          onChange: (key) => history.push(`/settings/${key}`),
         }}
-        onFinish={async (value) => {
-          const success = await handleUpdate(intl, value as API.Setting, Number(modalVisible));
-          if (success) {
-            setModalVisible(false);
-            if (actionRef.current) {
-              actionRef.current.reload();
+      >
+        <ProCard.TabPane key="user_settings" tab={<FormattedMessage id="user.settings" />}>
+          <UserSettings />
+        </ProCard.TabPane>
+
+        {Object.keys(configs).map((pkg) => (
+          <ProCard.TabPane
+            key={pkg}
+            tab={
+              <span>
+                <FormattedMessage id="package" /> {sanitizePackageName(pkg)}
+              </span>
             }
-          }
-        }}
-      ></SettingsModalForm>
+          >
+            <PackageForm values={configs} packageName={pkg} />
+            {pkg}
+          </ProCard.TabPane>
+        ))}
+      </ProCard>
     </PageContainer>
   );
 };
-
-export default TableList;
