+++ src/services/escola-lms/settings.ts
@@ -72,3 +72,14 @@ export async function settingGroups(options?: Record<string, any>) {
     ...(options || {}),
   });
 }
+
+
+export async function configs(
+  options?: Record<string, any>,
+) {
+  return request<API.ConfigsList>(`/api/admin/config`, {
+    method: 'GET',
+    /* useCache: true */ useCache: false,
+    ...(options || {}),
+  });
+}
